+++ contribs/locationchoice/src/main/java/org/matsim/contrib/locationchoice/bestresponse/LCPlan.java
@@ -18,7 +18,7 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.locationchoice.population;
+package org.matsim.contrib.locationchoice.bestresponse;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -29,11 +29,9 @@ import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Activity;
 import org.matsim.api.core.v01.population.Leg;
-import org.matsim.api.core.v01.population.Leg;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.api.core.v01.population.Plan;
 import org.matsim.api.core.v01.population.PlanElement;
-import org.matsim.api.core.v01.population.Plan;
 import org.matsim.api.core.v01.population.Route;
 import org.matsim.facilities.ActivityFacility;
 import org.matsim.utils.objectattributes.attributable.Attributes;
@@ -47,7 +45,7 @@ import org.matsim.utils.objectattributes.attributable.Attributes;
  * 
  * @author cdobler
  */
-public class LCPlan implements Plan {
+ class LCPlan implements Plan {
 
 	private Double score = null;
 	private Person person = null;
