+++ contribs/locationchoice/src/main/java/org/matsim/contrib/locationchoice/bestresponse/LCPlanElement.java
@@ -1,9 +1,10 @@
 /* *********************************************************************** *
  * project: org.matsim.*
+ * LCPlanElement.java
  *                                                                         *
  * *********************************************************************** *
  *                                                                         *
- * copyright       : (C) 2012 by the members listed in the COPYING,        *
+ * copyright       : (C) 2015 by the members listed in the COPYING,        *
  *                   LICENSE and WARRANTY file.                            *
  * email           : info at matsim dot org                                *
  *                                                                         *
@@ -17,41 +18,16 @@
  *                                                                         *
  * *********************************************************************** */
 
-package org.matsim.contrib.locationchoice.bestresponse.scoring;
+package org.matsim.contrib.locationchoice.bestresponse;
 
-import gnu.trove.map.TObjectDoubleMap;
-import gnu.trove.map.hash.TObjectDoubleHashMap;
+import org.matsim.api.core.v01.population.PlanElement;
 
-import java.util.Set;
+/**
+ * @author cdobler
+ */
+ interface LCPlanElement extends PlanElement {
 
-public class ScaleEpsilon {
+	public int getArrayIndex();
 	
-	private TObjectDoubleMap<String> epsilonFactors = new TObjectDoubleHashMap<>();
-	private boolean useSimpleTypes = false; // demand v1: e.g., s0.5, ... s24.0 = s
-	
-	public double getEpsilonFactor(String actType) {
-		if (this.useSimpleTypes) actType = actType.substring(0, 1);
-		return this.epsilonFactors.get(actType);
-	}
-	
-	public void setEpsilonFactor(String actType, double factor) {
-		this.epsilonFactors.put(actType, factor);
-	}
-	
-	public void setUseSimpleTypes(boolean useSimpleTypes) {
-		this.useSimpleTypes = useSimpleTypes;
-	}
-	
-	public boolean isFlexibleType(String actType) {
-		if (this.useSimpleTypes) actType = actType.substring(0, 1);
-		return this.epsilonFactors.containsKey(actType);
-	}
-	
-	public int getNumberOfFlexibleTypes() {
-		return this.epsilonFactors.size();
-	}
-	
-	public Set<String> getFlexibleTypes() {
-		return this.epsilonFactors.keySet();
-	}
+	public int getPlanElementIndex();
 }
