+++ contribs/locationchoice/src/main/java/org/matsim/contrib/locationchoice/timegeography/LocationChoicePlanStrategy.java
@@ -1,9 +1,10 @@
-package org.matsim.contrib.locationchoice;
+package org.matsim.contrib.locationchoice.timegeography;
 
 import org.matsim.api.core.v01.Scenario;
 import org.matsim.api.core.v01.population.HasPlansAndId;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.api.core.v01.population.Plan;
+import org.matsim.contrib.locationchoice.DestinationChoiceConfigGroup;
 import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.replanning.PlanStrategy;
 import org.matsim.core.replanning.PlanStrategyImpl;
